pyproject.toml
[project]
name = "ophthalmo-agent"
version = "0.1.0"
requires-python = ">=3.11"

dependencies = [
    "httpx>=0.27",
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "python-multipart>=0.0.9",

    # Extraction
    "PyMuPDF>=1.24",
    "Pillow>=10.0",
    "pytesseract>=0.3",

    # LLM
    "google-genai>=1.0",

    # Output
    "Jinja2>=3.1",
    "pandoc>=2.3",
    
    # Infra
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "tenacity>=8.0",
    "loguru>=0.7",
    "diskcache>=5.6",
    "typer>=0.12",
    "rich>=13.0",
    "fake-useragent>=1.5",

    # Cloud storage (Cloudflare R2 / S3-compatible) - optional for local dev
    "boto3>=1.34",
]

[project.optional-dependencies]
pdf-output = ["weasyprint>=62"]
dev = ["pytest>=8.0", "pytest-cov", "mypy", "ruff"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/download", "src/extract", "src/llm", "src/output", "src/utils", "src/config.py"]

[tool.ruff]
line-length = 88
target-version = "py311"
